Configuration: .build and .local when bdir is given interactively
Summary: will not affect existing configuration or analysis.

Until this commit, if the user gave the build directory interactively (had
not run './project configure' with '--build-dir'), the '.build' and
'.local' symbolic links were not created, resulting in a crash shortly
afterwards (when Maneage tried to write 'LOCAL.conf' in the build
directory!).

With this commit, the problem is solved by creating these links also when
the build directory is given interactively and after all the sanity checks
have passed.