README-hacking.md: customization for journals and published papers
Summary: will not affect existing configuration or analysis.

Until this commit, the customization checklist did not contain any
reference to the recently made branches of Maneage for customization on
certain journals (currently just the American Astronomical Society journals
and Astronomy and Astrophysics). Also, the list of papers published with
Maneage had not been updated for a long time.

With this commit, both have been added to 'README-hacking.md'.