apptainer.sh: more clear option names
Summary: The Apptainer script's option names have been renamed: '--sif' and
'--base-sif' are new names for the old '--project-name' and '--base-name'.

Until now the options that took SIF files in the Apptainer script were not
too clear (in their name): ending in '-name' (not clearly mentioning that
it needs a file name). Also, in the same script, we had forgot to add an
option to allow the user to choose a different base operating system!

With this commit, both the issues above have been fixed: the options end in
'-sif' (clearly informing the user that a SIF file is expected) and we now
have a new '--base-os' option. In the process, a few other parts of Maneage
were also polished:

  - README-hacking.md: the newly published J-PLUS PSF paper was added to
    the list of Maneage'd papers and the final journal information for the
    NASIM paper has been added. Some typos here and there were also fixed.

  - The curretnly used VIM needed the the '-std=gnu17': necessary now that
    the default Maneage GCC 15.2.0 for the high-level software which
    defaults to C23.