Configuration: Gnuastro not writing commit, versions, date or options
Summary: your project will only be affected if it uses Gnuastro and
depended on the FITS keywords that kept the information above (in the 0th
HDU).

Until now, there was no Maneage-specific configuration for Gnuastro, so as
in its default/manual operation, Gnuastro would keep all the metadata that
it keeps in manual/default mode. However, in a large pipeline that involves
the temporary creation of thousands of FITS files (which are usually
deleted shortly after), such metadata are just overhead. For the final
products of the pipeline, it is the responsibility of the pipeline designer
to only keep these in the final products of the pipeline, not all the
intermediate files.

With this commit, the default installation of Gnuastro in Maneage disables
all such metadata in its products in the output FITS files and provides
tips for users on the final metadata to include in their pipeline's
outputs.