Configuration: util-linux build issue fixed on macOS
Summary: this will not affect your software or analysis.

Until now, Maneage's configuration was failing during the build of
util-linux on macOS.

With this commit, the issue is solved by avoiding the building of bits.o
(the problematic part). It was checked both on GNU/Linux and macOS that the
removal of this util-linux feature does not affect any of the subsequent
programs that need util-linux.